Extract encryption work to separate folder
